move boxd to src/
